Highlight NPCs with already pending gifts.
If gifting a lot of NPCs in one sitting, it's likely to be hard to keep track of by memory, so this gives some slight UI treatment to visualize it (while still looking different from disabled NPCs, since their gifts _can_ still be swapped).

Updates the taste tooltips to include the item name, because it is confusing to have the icon for previously-sent item and a "dislikes this" message for the held item and makes it look like they dislike their sent gift instead. (It still might be slightly confusing because of the emoji; unsure if there's a perfect solution.)