Switch all view bindings to one-time.
The gift mail menu uses a one-shot view model with no dynamic UI, mainly because the menu needs to be closed in order to show the confirmation dialog, and/or choose a new item for the next gift, etc. Thus none of the models are INPC and the data is all static.